update input and dark mode functionality
