+++ index.scss
@@ -7,7 +7,7 @@
 :root {
   --border: 1px solid currentColor;
   --text-color: black;
-  --background-colour: white;
+  --background-colour: beige;
   --card-colour: springgreen;
 }
