+++ styles/inputs.scss
@@ -1,43 +1,59 @@
+@use 'sass:color';
+
 label {
   position: absolute;
   right: -2em;
   bottom: 50%;
 }
 
-.toggle,
-.toggle:before {
-  transition-property: background-color, transform, visibility;
-  transition-duration: 0.25s;
-  transition-timing-function: ease-in;
-}
-
-.toggle:before {
-  display: block;
-  position: absolute;
-  background: var(--background-color);
-  content: '';
-  width: 1em;
-  height: 1em;
-  border: 1px solid var(--text-color);
-  box-sizing: border-box;
-}
-
-.toggle:checked {
-  background-color: var(--text-color);
-
-  &:before {
-    transform: translateY(2em);
-  }
-}
-
-.toggle {
-  border: var(--border);
-  cursor: pointer;
-  position: relative;
-  width: 1em;
-  height: 3em;
+input[type='checkbox'] {
   -webkit-appearance: none;
   -moz-appearance: none;
   appearance: none;
   -webkit-tap-highlight-color: transparent;
+
+  box-shadow: inset 0 0 0 1px var(--text-color);
+  cursor: pointer;
+  position: relative;
+  width: 1.75em;
+  height: 1.75em;
+  background-color: var(--card-color);
+
+  &:before,
+  &:after {
+    content: '';
+    position: absolute;
+    border: 3px solid transparent;
+  }
+
+  &:before {
+    width: 0;
+    height: 100%;
+    right: 100%;
+    top: 0;
+    border-right-color: var(--text-color);
+    border-bottom-width: 0;
+  }
+
+  &:after {
+    width: 100%;
+    height: 0;
+    top: 100%;
+    right: 0;
+    border-top-color: var(--text-color);
+    border-left-width: 0;
+  }
+
+  &:checked {
+    transform: translate(-3px, 3px);
+    background-color: var(--text-color);
+    border: var(--border);
+    border-color: var(--text-color);
+    box-shadow: inset -3px 3px 0 0 var(--background-color);
+
+    &:after,
+    &:before {
+      display: none;
+    }
+  }
 }
